Cli: Improve error handling for installing plugins
